Extract common pre-execution test code
